wellhelpers: make templates private
to increase header encapsulation
